requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Werkzeug==2.3.7
Pillow==10.0.1
python-dotenv==1.0.0
google-cloud-aiplatform==1.38.1
google-auth==2.23.4
requests==2.31.0
google-cloud-storage==2.10.0
vertexai==1.38.1 